+++ driver/CMakeFiles/nytro-driver.dir/cmake_clean.cmake
@@ -1,0 +1,14 @@
+file(REMOVE_RECURSE
+  "../bin/nytro-driver"
+  "../bin/nytro-driver.pdb"
+  "CMakeFiles/nytro-driver.dir/link.d"
+  "CMakeFiles/nytro-driver.dir/src/config_loader.cpp.o"
+  "CMakeFiles/nytro-driver.dir/src/config_loader.cpp.o.d"
+  "CMakeFiles/nytro-driver.dir/src/driver.cpp.o"
+  "CMakeFiles/nytro-driver.dir/src/driver.cpp.o.d"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/nytro-driver.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()