+++ driver/CMakeFiles/nytro.dir/cmake_clean.cmake
@@ -1,0 +1,12 @@
+file(REMOVE_RECURSE
+  "../bin/nytro"
+  "../bin/nytro.pdb"
+  "CMakeFiles/nytro.dir/link.d"
+  "CMakeFiles/nytro.dir/driver.cpp.o"
+  "CMakeFiles/nytro.dir/driver.cpp.o.d"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/nytro.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()