+++ init.lua
@@ -1,0 +1,13 @@
+project = {
+	name = "test_app",
+	sources = {
+		"stress_test.ny",
+		"tests/test3.ny",
+	},
+	settings = {
+		verbose = true,
+		debug = true,
+		clean = true,
+		assembler = "nasm",
+	},
+}